+++ src/com/duggan/workflow/client/ui/events/DocumentSelectionEvent.java
@@ -3,7 +3,7 @@ package com.duggan.workflow.client.ui.events;
 import com.duggan.workflow.client.ui.util.DocMode;
 import com.google.gwt.event.shared.GwtEvent;
 import com.google.gwt.event.shared.EventHandler;
-import java.lang.Integer;
+import java.lang.Long;
 
 import javax.print.Doc;
 
@@ -13,18 +13,18 @@ public class DocumentSelectionEvent extends
 		GwtEvent<DocumentSelectionEvent.DocumentSelectionHandler> {
 
 	public static Type<DocumentSelectionHandler> TYPE = new Type<DocumentSelectionHandler>();
-	private Integer documentId;
+	private Long documentId;
 	private DocMode mode;
 
 	public interface DocumentSelectionHandler extends EventHandler {
 		void onDocumentSelection(DocumentSelectionEvent event);
 	}
 
-	public DocumentSelectionEvent(Integer documentId, DocMode docMode) {
+	public DocumentSelectionEvent(Long documentId, DocMode docMode) {
 		this.documentId = documentId;
 	}
 
-	public Integer getDocumentId() {
+	public Long getDocumentId() {
 		return documentId;
 	}
 	
@@ -46,7 +46,7 @@ public class DocumentSelectionEvent extends
 		return TYPE;
 	}
 
-	public static void fire(HasHandlers source, Integer documentId, DocMode docMode) {
+	public static void fire(HasHandlers source, Long documentId, DocMode docMode) {
 		source.fireEvent(new DocumentSelectionEvent(documentId, docMode));
 	}
 }
