+++ src/com/duggan/workflow/shared/requests/GetDocumentRequest.java
@@ -3,22 +3,20 @@ package com.duggan.workflow.shared.requests;
 import com.duggan.workflow.shared.responses.BaseResult;
 import com.duggan.workflow.shared.responses.GetDocumentResult;
 
-import java.lang.Integer;
-
 public class GetDocumentRequest extends BaseRequest<GetDocumentResult> {
 
-	private Integer id;
+	private Long id;
 
 	@SuppressWarnings("unused")
 	private GetDocumentRequest() {
 		// For serialization only
 	}
 
-	public GetDocumentRequest(Integer id) {
+	public GetDocumentRequest(Long id) {
 		this.id = id;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
